+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # TBD
 ### Fixed
-* Fail circus train if copy fails on tableReplicationStart
+* Fail circus train if copy fails on tableReplicationStart.  See [#52](https://github.com/HotelsDotCom/circus-train/issues/52).
 
 # 11.3.1 - 2018-03-21
 ### Changed